# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Support Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket detail with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket detail with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "404":
          description: Not found
  /tickets/{id}/status:
    patch:
      operationId: updateTicketStatus
      tags: [tickets]
      summary: Update ticket status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketStatusUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/reply:
    post:
      operationId: replyToTicket
      tags: [tickets]
      summary: Reply to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tickets/{id}/notes:
    post:
      operationId: addTicketNote
      tags: [tickets]
      summary: Add an internal note to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketNoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── Branch ────────────────────────────────────────────────────────────────
...
    TicketStatus:
      type: string
      enum:
        - open
        - in_progress
        - waiting_client
        - waiting_analyst
        - closed
    Ticket:
      type: object
      properties:
        id:
          type: integer
        ticketNumber:
          type: string
        whatsappPhone:
          type: string
        clientName:
          type: string
          nullable: true
        branchId:
          type: integer
          nullable: true
        branchName:
          type: string
          nullable: true
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        description:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        assignedTo:
          type: integer
          nullable: true
        assigneeName:
          type: string
          nullable: true
        reopenCount:
          type: integer
        firstResponseAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        lastMessageAt:
          type: string
          nullable: true
        botMode:
          type: string
          nullable: true
      required:
        - id
        - ticketNumber
        - whatsappPhone
        - description
        - status
        - createdAt
        - updatedAt
    TicketMessage:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        direction:
          type: string
        type:
          type: string
        content:
          type: string
        mediaUrl:
          type: string
          nullable: true
        senderName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - direction
        - type
        - content
        - createdAt
    TicketDetail:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/Ticket"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
      required:
        - ticket
        - messages
    TicketStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum:
            - open
            - in_progress
            - waiting_client
            - waiting_analyst
            - closed
        note:
          type: string
      required:
        - status
    TicketReplyInput:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    TicketNoteInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content